+++ packages/ricos-content/web/src/converters/draft/fromDraft/fromDraft.ts
@@ -7,6 +7,7 @@ import { generateId } from '../../generateRandomId';
 import { getTextNodes } from './getTextNodes';
 import { getEntity, getNodeStyle, getTextStyle } from './getRicosEntityData';
 import { createParagraphNode, initializeMetadata } from '../../nodeUtils';
+import { nestedNodesConverters } from './nestedNodesUtils';
 
 export const ensureRicosContent = (content: RichContent | DraftContent): RichContent =>
   'blocks' in content ? fromDraft(content) : content;
@@ -58,28 +59,6 @@ export const fromDraft = (draftJSON: DraftContent): RichContent => {
       }
     }
   };
-
-  const parseCollapsiblePairs = (entity): Node | undefined =>
-    entity?.collapsibleListData?.pairs?.map(
-      (pair: { key: string; title: DraftContent; content: DraftContent }) => ({
-        id: pair.key,
-        type: Node_Type.COLLAPSIBLE_ITEM,
-        nodes: [
-          {
-            id: generateId(),
-            type: Node_Type.COLLAPSIBLE_ITEM_TITLE,
-            nodes: fromDraft(pair.title).nodes,
-          },
-          {
-            id: generateId(),
-            type: Node_Type.COLLAPSIBLE_ITEM_BODY,
-            nodes: fromDraft(pair.content).nodes,
-          },
-        ],
-      })
-    );
-
-  const nestedNodesConverters = { [Node_Type.COLLAPSIBLE_LIST]: parseCollapsiblePairs };
 
   const parseAtomicBlock = ({ key, data, entityRanges }: RicosContentBlock): Node | null => {
     if (entityRanges && entityRanges.length) {