+++ packages/ricos-content/web/src/converters/plainText/toPlainText/complexPlainText.ts
@@ -74,4 +74,7 @@ Collapsible List
 
 #LivingTheDream
 Yo yo yo!
+
+
+
 `;